requirements.txt
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.8.2
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
